@@ -18,10 +18,11 @@ package co.paulburke.android.itemtouchhelperdemo;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.app.ListFragment;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
+
+import androidx.fragment.app.ListFragment;
 
 /**
  * @author Paul Burke (ipaulpro)
